backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.0
SQLAlchemy==2.0.38
alembic==1.14.1
asyncpg==0.30.0
python-jose==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
email-validator==2.3.0
redis==5.2.1
celery==5.4.0
sqladmin==0.20.1
prometheus-fastapi-instrumentator==7.0.2
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
